Split tests out
